+++ src/bundle/Resources/public/js/CKEditor/custom-tags/helpers/panel-helper.js
@@ -1,0 +1,17 @@
+const setPanelContentMaxHeight = (balloonView) => {
+    const { innerHeight: windowHeight } = window;
+    const { top: panelTopPosition, element: panelNode } = balloonView;
+    const panelHeader = panelNode.querySelector('.ibexa-custom-tag-panel-header');
+    const panelContent = panelNode.querySelector('.ibexa-custom-tag-panel-content');
+    const panelFooter = panelNode.querySelector('.ibexa-custom-tag-panel-footer');
+    const panelHeaderHeight = panelHeader?.offsetHeight ?? 0;
+    const panelFooterHeight = panelFooter?.offsetHeight ?? 0;
+    const isPanelOverTopWindowEdge = panelTopPosition < 0;
+    const maxHeightValue = isPanelOverTopWindowEdge
+        ? panelContent.offsetHeight - Math.abs(panelTopPosition)
+        : windowHeight - panelTopPosition - panelHeaderHeight - panelFooterHeight;
+
+    panelContent.style.maxHeight = `${maxHeightValue}px`;
+};
+
+export { setPanelContentMaxHeight };
